requirements.txt
garminconnect>=0.2.20
python-dotenv>=1.0.0
pandas>=2.2.0
SQLAlchemy>=2.0.0
tqdm>=4.66.0
streamlit>=1.35.0
plotly>=5.20.0
bcrypt>=4.0.0
cryptography>=42.0.0
